+++ cmd/app/main.go
@@ -2,9 +2,10 @@ package main
 
 import (
 	"fmt"
+	"github.com/rs/zerolog/log"
 	"go-microservice-starter/cmd/app/config"
 	"go-microservice-starter/cmd/app/routes"
-	"log"
+	"go-microservice-starter/internal/database"
 	"net/http"
 )
 
@@ -16,21 +17,40 @@ func main() {
 }
 
 func start() error {
-	log.Println("initializing application...")
+	log.Info().Msg("initializing application...")
 
-	// set up service
+	// init service
 	config.Set()
 	port := config.Registry.GetString("SERVER_PORT")
 
-	// set up routes
+	// init db
+	dbSettings := config.GetDBSettings()
+	db, err := database.NewConnection(dbSettings)
+	if err != nil {
+		return err
+	}
+
+	// init services
+	vs := config.NewValidateService()
+	hs := config.NewHealthService(db)
+
+	// init routes
+	routeServices := routes.Services{
+		ValidatorService: vs,
+		HelloService:     hs,
+	}
 	router := routes.NewRouter()
-	router.Setup()
+	err = router.Setup(routeServices)
+	if err != nil {
+		log.Error().Msg("error setting up routes")
+		return err
+	}
 
 	// start application
-	log.Printf("service running on port %s\n", port)
-	err := http.ListenAndServe(port, router.Router)
+	log.Info().Msg(fmt.Sprintf("service running on port %s", port))
+	err = http.ListenAndServe(port, router.Router)
 	if err != nil {
-		log.Println("error setting up application")
+		log.Error().Msg("error setting up application")
 		return err
 	}