+++ cmd/app/routes/routes.go
@@ -1,7 +1,10 @@
 package routes
 
 import (
-	"log"
+	"fmt"
+	"github.com/rs/zerolog/log"
+	"go-microservice-starter/internal/hello"
+	"go-microservice-starter/internal/validate"
 	"net/http"
 
 	"github.com/gorilla/mux"
@@ -15,15 +18,43 @@ const (
 	_v1  = "/v1"
 )
 
+type Services struct {
+	ValidatorService validate.Service
+	HelloService     hello.Service
+}
+
+type handler struct {
+	Services
+}
+
+// newHandler returns new handler
+func newHandler(services Services) (handler, error) {
+	if services.ValidatorService == nil {
+		return handler{}, fmt.Errorf("no validator service provided")
+	}
+	if services.HelloService == nil {
+		return handler{}, fmt.Errorf("no hello service provided")
+	}
+
+	return handler{services}, nil
+}
+
 // NewRouter returns Router pointer
 func NewRouter() *Router {
 	return &Router{}
 }
 
 // Setup creates routes for the app
-func (r *Router) Setup() {
-	log.Println("initializing routes...")
+func (r *Router) Setup(services Services) error {
+	log.Info().Msg("initializing routes...")
+
+	h, err := newHandler(services)
+	if err != nil {
+		return err
+	}
 
 	r.Router = mux.NewRouter()
-	r.Router.HandleFunc(_api+_v1+"/health", getHealth).Methods(http.MethodGet)
+	r.Router.HandleFunc(_api+_v1+"/hello", h.postHello).Methods(http.MethodPost)
+
+	return nil
 }