+++ database/query/query.sql
@@ -1,0 +1,9 @@
+-- name: GetUsers :many
+select
+    `id_user`,
+    `username`,
+    `email`,
+    `created_on`,
+    `last_login`
+from
+    users;