+++ internal/hello/sayhello.go
@@ -1,0 +1,14 @@
+package hello
+
+import (
+	"context"
+	"fmt"
+)
+
+// SayHello is just an example of some business logic
+func (s *service) SayHello(_ context.Context, r PostRequest) PostResponse {
+	return PostResponse{
+		Status:  "ok",
+		Message: fmt.Sprintf("hello, %s", r.Name),
+	}
+}