+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.0] - 2022-05-29
+### Added
+- implemented service
+
 ## [0.0.1] - 2021-10-13
 ### Added
-- Initialized service
+- initialized service